s/scalahost/semanticdb-scalac/g
