+++ .scalafmt.conf
@@ -7,15 +7,15 @@ maxColumn = 100
 optIn.annotationNewlines = true
 project.git = true
 project.includeFilters = [
-  scalahost
+  semanticdb-scalac
   scalameta/contrib
 ]
 project.excludeFilters = [
   sbt-test
   contrib/package.scala
-  /scalahost/nsc/src/main/resources/scalac-plugin.xml
-  /scalahost/nsc/src/main/scala-2.11.11/scala/meta/internal/scalahost/compilers/ScalahostAnalyzer.scala
-  /scalahost/nsc/src/main/scala-2.12.3/scala/meta/internal/scalahost/compilers/ScalahostAnalyzer.scala
+  scalameta/semanticdb-scalac/src/main/resources/scalac-plugin.xml
+  scalameta/semanticdb-scalac/src/main/scala-2.11.11/scala/meta/internal/SemanticdbAnalyzer.scala
+  scalameta/semanticdb-scalac/src/main/scala-2.12.3/scala/meta/internal/SemanticdbAnalyzer.scala
 ]
 runner.dialect = Scala211
 runner.optimizer.forceConfigStyleOnOffset = -1
