+++ CONTRIBUTING.md
@@ -3,7 +3,7 @@
 - **Bugs**:
   - Contain steps to reproduce.
   - Contain a code sample that exhibits the error.
-  - *Ideally* contain your scalameta/scalahost/paradise versions.
+  - *Ideally* contain your scalameta/semanticdb versions.
   - *Ideally* contain the scala version you are using.
   - Inform us if the issue is blocking you with no visible workaround
     - This will give the ticket priority
