--- scalahost/nsc/src/main/scala/scala/meta/internal/scalahost/ScalahostPlugin.scala
@@ -1,44 +1,0 @@
-package scala.meta.internal
-package scalahost
-
-import scala.meta.internal.semanticdb.FailureMode
-import scala.meta.internal.semanticdb.SemanticdbMode
-import scala.meta.io.AbsolutePath
-import scala.tools.nsc.Global
-import scala.tools.nsc.plugins.{Plugin, PluginComponent}
-
-class ScalahostPlugin(val global: Global)
-    extends Plugin
-    with HijackAnalyzer
-    with HijackReporter
-    with ScalahostPipeline {
-  val name = ScalahostPlugin.name
-  val description = "scala.meta's connector to the Scala compiler"
-  hijackAnalyzer()
-  hijackReporter()
-  val components = List[PluginComponent](ScalahostComponent)
-
-  override def init(options: List[String], error: (String) => Unit): Boolean = {
-    def err(msg: String): Unit = {
-      g.reporter.error(g.NoPosition, s"[scalahost] $msg")
-    }
-    options.foreach {
-      case SetSourceroot(path) =>
-        val abspath = AbsolutePath(path)
-        config.setSourceroot(abspath)
-      case SetSemanticdb(SemanticdbMode(mode)) =>
-        config.setSemanticdbMode(mode)
-      case SetFailures(FailureMode(severity)) =>
-        config.setFailures(severity)
-      case SetSemanticdb(els) =>
-        err(s"Unknown semanticdb $els. Expected one of: ${SemanticdbMode.all.mkString(", ")} ")
-      case els =>
-        err(s"Ignoring unknown option $els")
-    }
-    true
-  }
-}
-
-object ScalahostPlugin {
-  val name = "scalahost"
-}
