--- scalahost/nsc/src/main/scala/scala/meta/internal/scalahost/ScalahostReporter.scala
@@ -1,22 +1,0 @@
-package scala.meta.internal
-package scalahost
-
-import scala.reflect.internal.util.Position
-import scala.tools.nsc.reporters.Reporter
-import scala.tools.nsc.reporters.StoreReporter
-
-class ScalahostReporter(underlying: Reporter) extends StoreReporter {
-  override protected def info0(
-      pos: Position,
-      msg: String,
-      severity: Severity,
-      force: Boolean): Unit = {
-    super.info0(pos, msg, severity, force)
-    severity.id match {
-      case 0 => underlying.info(pos, msg, force)
-      case 1 => underlying.warning(pos, msg)
-      case 2 => underlying.error(pos, msg)
-      case _ =>
-    }
-  }
-}
