+++ scalameta/semanticdb-integration/src/main/scala/example/Sugar.scala
