+++ scalameta/semanticdb-scalac/src/main/resources/scalac-plugin.xml
@@ -1,0 +1,4 @@
+<plugin>
+  <name>semanticdb</name>
+  <classname>scala.meta.internal.SemanticdbPlugin</classname>
+</plugin>
