+++ scalameta/semanticdb-scalac/src/main/scala/scala/meta/internal/ReflectionToolkit.scala
