+++ scalameta/semanticdb-scalac/src/main/scala/scala/meta/internal/semanticdb/ReporterOps.scala
