+++ scalameta/semanticdb-scalac/src/main/scala/scala/meta/internal/semanticdb/SymbolOps.scala
